chore: update gen scripts (#1611)
- rename global scripts to match better it functionalities
- remove duplicate dts generation
