+++ .prettierignore
@@ -1,4 +1,5 @@
 # prevent formatting for generated files
 CHANGELOG.md
-packages/oruga/src/components/types.ts
+packages/oruga/src/config.d.ts
+packages/oruga/src/globals.d.ts
 **/theme-*.md
