+++ eslint.config.mjs
@@ -22,7 +22,8 @@ export default [
     "*.d.ts",
     // prevent lint for generated files
     "CHANGELOG.md",
-    "**/packages/oruga/src/components/types.ts",
+    "**/packages/oruga/src/config.d.ts",
+    "**/packages/oruga/src/globals.d.ts",
     "**/theme-*.md",
   ]),
