+++ package.json
@@ -31,8 +31,8 @@
     "test:coverage": "npm run test:coverage --workspace @oruga-ui/oruga-next",
     "test:ts": "npm run test:ts --workspace @oruga-ui/oruga-next",
     "gen": "run-s gen:*",
-    "gen:volar": "tsx .scripts/gen-volar-dts.ts",
-    "gen:types": "tsx .scripts/gen-comp-types.ts",
+    "gen:volar": "tsx .scripts/gen-globals-dts.ts",
+    "gen:types": "tsx .scripts/gen-config-dts.ts",
     "gen:docs": "npm run docs:gen --workspace @oruga-ui/docs-next",
     "version": "tsx .scripts/sync-version.ts && npm i",
     "changelog": "conventional-changelog -p conventionalcommits -i CHANGELOG.md -s",
