--- packages/oruga/components.d.ts
@@ -1,51 +1,0 @@
-// Auto generated component declarations
-declare module "vue" {
-    export interface GlobalComponents {
-        OAutocomplete: (typeof import("../oruga/src/components"))["OAutocomplete"];
-        OBreadcrumb: (typeof import("../oruga/src/components"))["OBreadcrumb"];
-        OBreadcrumbItem: (typeof import("../oruga/src/components"))["OBreadcrumbItem"];
-        OButton: (typeof import("../oruga/src/components"))["OButton"];
-        OCarousel: (typeof import("../oruga/src/components"))["OCarousel"];
-        OCarouselItem: (typeof import("../oruga/src/components"))["OCarouselItem"];
-        OCheckbox: (typeof import("../oruga/src/components"))["OCheckbox"];
-        OCollapse: (typeof import("../oruga/src/components"))["OCollapse"];
-        ODatepicker: (typeof import("../oruga/src/components"))["ODatepicker"];
-        ODatetimepicker: (typeof import("../oruga/src/components"))["ODatetimepicker"];
-        ODialog: (typeof import("../oruga/src/components"))["ODialog"];
-        ODropdown: (typeof import("../oruga/src/components"))["ODropdown"];
-        ODropdownItem: (typeof import("../oruga/src/components"))["ODropdownItem"];
-        OField: (typeof import("../oruga/src/components"))["OField"];
-        OIcon: (typeof import("../oruga/src/components"))["OIcon"];
-        OInput: (typeof import("../oruga/src/components"))["OInput"];
-        OListbox: (typeof import("../oruga/src/components"))["OListbox"];
-        OListItem: (typeof import("../oruga/src/components"))["OListItem"];
-        OLoading: (typeof import("../oruga/src/components"))["OLoading"];
-        OMenu: (typeof import("../oruga/src/components"))["OMenu"];
-        OMenuItem: (typeof import("../oruga/src/components"))["OMenuItem"];
-        OModal: (typeof import("../oruga/src/components"))["OModal"];
-        ONotification: (typeof import("../oruga/src/components"))["ONotification"];
-        ONotificationNotice: (typeof import("../oruga/src/components"))["ONotificationNotice"];
-        OPagination: (typeof import("../oruga/src/components"))["OPagination"];
-        ORadio: (typeof import("../oruga/src/components"))["ORadio"];
-        OSelect: (typeof import("../oruga/src/components"))["OSelect"];
-        OSidebar: (typeof import("../oruga/src/components"))["OSidebar"];
-        OSkeleton: (typeof import("../oruga/src/components"))["OSkeleton"];
-        OSlider: (typeof import("../oruga/src/components"))["OSlider"];
-        OSliderTick: (typeof import("../oruga/src/components"))["OSliderTick"];
-        OStepItem: (typeof import("../oruga/src/components"))["OStepItem"];
-        OSteps: (typeof import("../oruga/src/components"))["OSteps"];
-        OSwitch: (typeof import("../oruga/src/components"))["OSwitch"];
-        OTabItem: (typeof import("../oruga/src/components"))["OTabItem"];
-        OTable: (typeof import("../oruga/src/components"))["OTable"];
-        OTableColumn: (typeof import("../oruga/src/components"))["OTableColumn"];
-        OTabs: (typeof import("../oruga/src/components"))["OTabs"];
-        OTag: (typeof import("../oruga/src/components"))["OTag"];
-        OTaginput: (typeof import("../oruga/src/components"))["OTaginput"];
-        OTimepicker: (typeof import("../oruga/src/components"))["OTimepicker"];
-        OTooltip: (typeof import("../oruga/src/components"))["OTooltip"];
-        OTree: (typeof import("../oruga/src/components"))["OTree"];
-        OTreeItem: (typeof import("../oruga/src/components"))["OTreeItem"];
-        OUpload: (typeof import("../oruga/src/components"))["OUpload"];
-    }
-}
-export {};
