+++ packages/oruga/vite.config.ts
@@ -22,6 +22,7 @@ export default defineConfig({
             entryRoot: "./src",
             outDir: "./dist/types",
             bundledPackages: ["vue-component-type-helpers"],
+            copyDtsFiles: true,
         }),
     ],
     resolve: {
